+++ frontend/src/pages/UserProfile.jsx
@@ -8,6 +8,7 @@ import Modal from "../components/common/Modal";
 import { jwtDecode } from 'jwt-decode';
 import EditUserModal from "../components/modals/EditUserModal";
 import ReusableTable from "../components/common/ReusableTable";
+import UserFoodOrdersList from "../components/lists/UserFoodOrdersList";
 
 const UserProfile = () => {
   const { userId } = useParams();
@@ -78,76 +79,6 @@ const UserProfile = () => {
       .finally(() => setUserOrdersLoading(false));
   }, [userId]);
 
-
-
-  // Table columns for user food orders
-  const userOrderColumns = [
-    {
-      key: "orderId",
-      header: "Order ID",
-      sortable: true,
-    },
-    {
-      key: "orderedAt",
-      header: "Ordered At",
-      sortable: true,
-      render: (val) => val ? new Date(val).toLocaleString() : "-",
-    },
-    {
-      key: "expectTime",
-      header: "Expect Time",
-      sortable: true,
-      render: (val) => val ? new Date(val).toLocaleString() : "-",
-    },
-    
-    {
-      key: "status",
-      header: "Status",
-      sortable: true,
-      render: (val) => {
-        let style = "";
-        if (val === "Delivered") {
-          style = "bg-green-500 text-white";
-        } else if (val === "Cancelled") {
-          style = "bg-red-500 text-white";
-        } else if (val === "Cancelled by User") {
-          style = "bg-yellow-500 text-white";
-        } else if (val === "Pending") {
-          style = "bg-blue-500 text-white";
-        } else if (val === "Preparing") {
-          style = "bg-indigo-500 text-white";
-        }
-        return (
-          <span className={`px-3 py-1 rounded font-semibold text-xs shadow ${style}`}>
-            {val}
-          </span>
-        );
-      },
-    },
-    {
-      key: "totalPrice",
-      header: "Total Price",
-      sortable: true,
-      render: (val) => `${val} LKR`,
-    },
-    {
-      key: "items",
-      header: "Items",
-      render: (items) =>
-        <ul className="list-disc ml-4">
-          {items.map((item, idx) => (
-            <li key={idx}>
-              <span className="font-mono">{item.foodCode}</span> - {item.foodId?.name || item.name} ({item.quantity} x {item.price} LKR)
-            </li>
-          ))}
-        </ul>,
-    },
-    {
-      key: "specialRequest",
-      header: "Special Request",
-      render: (val) => val || "-",
-    },
-  ];
 
   // Helpers
   const getCompanyName = (companyId) => {
@@ -314,28 +245,15 @@ const UserProfile = () => {
             ))}
           </div>
         </div>
-        {/* Pass selectedRoomId as prop to ACList */}
+
+        
         <ACList userId={userId} selectedRoomId={selectedRoomId} />
         <DoorList userId={userId} selectedRoomId={selectedRoomId} />
         <LightList userId={userId} selectedRoomId={selectedRoomId} />
+        <UserFoodOrdersList userOrders={userOrders} loading={userOrdersLoading}/>
+
       </div>
 
-      {/* User Food Orders Section */}
-      <div className="mt-10">
-        <h3 className="text-xl font-bold mb-4">User Food Orders</h3>
-        <ReusableTable
-          columns={userOrderColumns}
-          data={userOrders}
-          searchable={true}
-          sortable={true}
-          pagination={true}
-          pageSize={10}
-          striped={true}
-          hover={true}
-          loading={userOrdersLoading}
-          emptyMessage="No food orders found for this user."
-        />
-      </div>
 
       <EditUserModal
         isVisible={showEditModal}
